Created controllers, services and tests

+++ src/main/java/fr/motormingle/api/repository/CarRepository.java
@@ -4,5 +4,5 @@ import fr.motormingle.api.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CarRepository extends JpaRepository<Car, Integer>, JpaSpecificationExecutor<Car> {
+public interface CarRepository extends JpaRepository<Car, Long>, JpaSpecificationExecutor<Car> {
 }
