+++ src/main/java/fr/motormingle/api/repository/MotorbikeRepository.java
@@ -4,5 +4,5 @@ import fr.motormingle.api.entity.Motorbike;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface MotorbikeRepository extends JpaRepository<Motorbike, Integer>, JpaSpecificationExecutor<Motorbike> {
+public interface MotorbikeRepository extends JpaRepository<Motorbike, Long>, JpaSpecificationExecutor<Motorbike> {
 }
